chore: consolidated how the allow-list was being retrieved/persisted
